Switch all Legion fields to float32.

+++ spinifel/legion/autocorrelation.py
@@ -25,7 +25,7 @@ def gen_random_orientations(orientations, N_images_per_rank):
 @nvtx.annotate("legion/autocorrelation.py", is_prefix=True)
 def get_random_orientations():
     N_images_per_rank = parms.N_images_per_rank
-    fields_dict = {"quaternions": pygion.float64}
+    fields_dict = {"quaternions": pygion.float32}
     sec_shape = (4,)
     orientations, orientations_p = lgutils.create_distributed_region(
         N_images_per_rank, fields_dict, sec_shape)
@@ -80,8 +80,8 @@ def gen_nonuniform_positions(orientations, nonuniform, pixel_position):
 @nvtx.annotate("legion/autocorrelation.py", is_prefix=True)
 def get_nonuniform_positions(orientations, orientations_p, pixel_position):
     N_images_per_rank = parms.N_images_per_rank
-    fields_dict = {"H": pygion.float64, "K": pygion.float64,
-                   "L": pygion.float64}
+    fields_dict = {"H": pygion.float32, "K": pygion.float32,
+                   "L": pygion.float32}
     sec_shape = parms.reduced_det_shape
     nonuniform, nonuniform_p = lgutils.create_distributed_region(
         N_images_per_rank, fields_dict, sec_shape)
@@ -183,7 +183,7 @@ def prepare_solve(slices, slices_p, nonuniform, nonuniform_p,
     nonuniform_v, nonuniform_v_p = get_nonuniform_positions_v(
         nonuniform, nonuniform_p, reciprocal_extent)
     uregion = Region((M,)*3,
-                     {"ADb": pygion.float64, "F_antisupport": pygion.float32})
+                     {"ADb": pygion.float32, "F_antisupport": pygion.float32})
     uregion_ups = Region((M_ups,)*3, {"F_conv_": pygion.complex128})
     prep_Fconv(uregion_ups, nonuniform_v, nonuniform_v_p,
                ac, weights, M_ups, Mtot, N,
@@ -354,7 +354,7 @@ def solve_ac(generation,
 
     #N_ranks = 5
     N_procs = Tunable.select(Tunable.GLOBAL_PYS).get()
-    results = Region((N_procs * M, M, M), {"ac": pygion.float64})
+    results = Region((N_procs * M, M, M), {"ac": pygion.float32})
     results_p = Partition.restrict(results, (N_procs,), [[M], [0], [0]], [M, M, M])
 
     alambda = 1
